Move Foundation.WebApi to new AspNetCoreUtils project
