+++ account-management/AccountManagement.sln
@@ -2,8 +2,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{EEA1463E-71F8-44E6-8055-787DAA5E62D2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Foundation", "src\Foundation\Foundation.csproj", "{92CECDBD-ADDD-437B-A848-9A663BA56213}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "src\Domain\Domain.csproj", "{254BF290-0F52-40FD-9776-25660C34D8CE}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "src\Application\Application.csproj", "{51EA2C15-ACB4-40FF-A6F2-05512BED7264}"
@@ -24,16 +22,14 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DomainModeling", "..\backen
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PersistenceInfrastructure", "..\backend-foundation\src\PersistenceInfrastructure\PersistenceInfrastructure.csproj", "{3AD09E47-0F64-4240-80C1-58940B37DCAC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspNetCoreUtils", "..\backend-foundation\src\AspNetCoreUtils\AspNetCoreUtils.csproj", "{FCCDC977-AA69-4D46-8AD9-586693E53CE0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{92CECDBD-ADDD-437B-A848-9A663BA56213}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{92CECDBD-ADDD-437B-A848-9A663BA56213}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{92CECDBD-ADDD-437B-A848-9A663BA56213}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{92CECDBD-ADDD-437B-A848-9A663BA56213}.Release|Any CPU.Build.0 = Release|Any CPU
 		{254BF290-0F52-40FD-9776-25660C34D8CE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{254BF290-0F52-40FD-9776-25660C34D8CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{254BF290-0F52-40FD-9776-25660C34D8CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -66,9 +62,12 @@ Global
 		{3AD09E47-0F64-4240-80C1-58940B37DCAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3AD09E47-0F64-4240-80C1-58940B37DCAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3AD09E47-0F64-4240-80C1-58940B37DCAC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FCCDC977-AA69-4D46-8AD9-586693E53CE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FCCDC977-AA69-4D46-8AD9-586693E53CE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FCCDC977-AA69-4D46-8AD9-586693E53CE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FCCDC977-AA69-4D46-8AD9-586693E53CE0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{92CECDBD-ADDD-437B-A848-9A663BA56213} = {EEA1463E-71F8-44E6-8055-787DAA5E62D2}
 		{254BF290-0F52-40FD-9776-25660C34D8CE} = {EEA1463E-71F8-44E6-8055-787DAA5E62D2}
 		{51EA2C15-ACB4-40FF-A6F2-05512BED7264} = {EEA1463E-71F8-44E6-8055-787DAA5E62D2}
 		{E4D8E1A7-0712-4233-A231-6F7AFB7AD626} = {EEA1463E-71F8-44E6-8055-787DAA5E62D2}
@@ -77,5 +76,6 @@ Global
 		{B1E9935F-2A8C-41F1-BC72-BB5D062F9CB0} = {0E389279-C3EB-4F9F-8068-8A12F0D526F2}
 		{4D61C57C-33EB-4A9F-BA35-80F4F1AAEAA2} = {F01E4DC8-2A8B-4CB9-893A-B3B8FF2EFE22}
 		{3AD09E47-0F64-4240-80C1-58940B37DCAC} = {F01E4DC8-2A8B-4CB9-893A-B3B8FF2EFE22}
+		{FCCDC977-AA69-4D46-8AD9-586693E53CE0} = {F01E4DC8-2A8B-4CB9-893A-B3B8FF2EFE22}
 	EndGlobalSection
 EndGlobal
