+++ account-management/src/Infrastructure/Infrastructure.csproj
@@ -9,6 +9,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <ProjectReference Include="..\..\..\backend-foundation\src\PersistenceInfrastructure\PersistenceInfrastructure.csproj"/>
         <ProjectReference Include="..\Application\Application.csproj"/>
     </ItemGroup>
