+++ account-management/tests/Foundation.Tests/Foundation.Tests.csproj
@@ -12,10 +12,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\src\Foundation\Foundation.csproj"/>
-    </ItemGroup>
-
-    <ItemGroup>
         <PackageReference Include="FluentAssertions" Version="6.10.0"/>
         <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="7.0.4"/>
         <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="7.0.5"/>
@@ -31,6 +27,12 @@
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
             <PrivateAssets>all</PrivateAssets>
         </PackageReference>
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\..\..\backend-foundation\src\AspNetCoreUtils\AspNetCoreUtils.csproj"/>
+        <ProjectReference Include="..\..\..\backend-foundation\src\DomainModeling\DomainModeling.csproj"/>
+        <ProjectReference Include="..\..\..\backend-foundation\src\PersistenceInfrastructure\PersistenceInfrastructure.csproj"/>
     </ItemGroup>
 
 </Project>
