+++ backend-foundation/src/AspNetCoreUtils/AspNetCoreUtils.csproj
@@ -2,10 +2,10 @@
 
     <PropertyGroup>
         <TargetFramework>net7.0</TargetFramework>
-        <AssemblyName>PlatformPlatform.Foundation</AssemblyName>
-        <RootNamespace>PlatformPlatform.Foundation</RootNamespace>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <AssemblyName>PlatformPlatform.Foundation.AspNetCoreUtils</AssemblyName>
+        <RootNamespace>PlatformPlatform.Foundation.AspNetCoreUtils</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
@@ -13,16 +13,8 @@
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="IdGen" Version="3.0.3"/>
         <PackageReference Include="JetBrains.Annotations" Version="2022.3.1"/>
-        <PackageReference Include="MediatR" Version="12.0.1"/>
         <PackageReference Include="Microsoft.AspNetCore.Mvc.ApiExplorer" Version="2.2.0"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.4"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.4"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="7.0.4">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
     </ItemGroup>
 
     <ItemGroup>
