+++ backend-foundation/src/DomainModeling/DomainModeling.csproj
@@ -12,6 +12,6 @@
         <PackageReference Include="IdGen" Version="3.0.3"/>
         <PackageReference Include="JetBrains.Annotations" Version="2022.3.1"/>
         <PackageReference Include="MediatR" Version="12.0.1"/>
+        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.0"/>
     </ItemGroup>
-
 </Project>
